Backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-chess==1.999
pydantic==1.10.9
python-dotenv
redis==3.5.3